factor out common HRRR template config
